pyproject.toml
[project]
name = "free-tg-mtproxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncssh>=2.18.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
]
